metabase/processador/requirements.txt
paho-mqtt==1.5.1
psycopg2-binary==2.8.6